Update Contributors.md
I think it is about time to make an update on this one.

Sorry for my long absence and lack of communication. I'm currently super busy pursuing my own PhD and barely have time for something outside of it.
I hope you're all doing well and wish you all the best with continuing on this project, maybe some day I will have time to come back to it.